Working on improvements. Balance is now displayed on the top. In place of older balance I want to place calculations (selected nominal * counter) and then after accepting it we wold add it to the array. This would fix the issue with incorrect output and simplify the code.

+++ Kasjer/Models/ItemModel.swift
@@ -10,7 +10,7 @@ import SwiftUI
 
 struct ItemModel: Identifiable {
     var nominal: Double
-    var count: Int
+    var count: Double
 //    let color: Color
     let id: UUID
     let theValue: Double
